Move test to remote settings file (since it uses kinto).
