+++ app/tests/integration/nimbus/test_remote_settings.py
@@ -1,4 +1,7 @@
+from urllib.parse import urljoin
+
 import pytest
+from nimbus.pages.experimenter.home import HomePage
 from nimbus.pages.experimenter.summary import SummaryPage
 
 
@@ -99,3 +102,19 @@ def test_takeaways(
 
     assert summary.takeaways_summary_text == expected_summary
     assert summary.takeaways_recommendation_badge_text == "Change course"
+
+
+@pytest.mark.remote_settings
+@pytest.mark.xdist_group(name="group2")
+def test_check_live_status_on_home_page(
+    selenium, base_url, create_experiment, kinto_client, experiment_name, slugify
+):
+    experiment_slug = str(slugify(experiment_name))
+
+    summary = create_experiment(selenium)
+    summary.launch_and_approve()
+    kinto_client.approve()
+    summary = SummaryPage(selenium, urljoin(base_url, experiment_slug)).open()
+    summary.wait_for_live_status()
+    home = HomePage(selenium, base_url).open()
+    assert True in [experiment_name in item.text for item in home.tables[0].experiments]
